+++ registry/lib/ui/default_components/label.dart
@@ -2,50 +2,64 @@
 
 import 'package:fcnui_base/fcnui_base.dart';
 import 'package:flutter/material.dart';
+import 'package:registry/ui/default_components/fcnui_theme.dart';
 import 'disabled.dart';
 
-class LabelModel extends Equatable {
-  final String text;
-  final bool isRequired;
-  final bool enabled;
-  const LabelModel({
-    required this.text,
-    this.enabled = true,
-    this.isRequired = false,
-  });
+class LabelDecoration extends DecorationImpl {
+  LabelDecoration(
+    super.context, {
+    required LabelChild child,
+    LabelState? state,
+  }) {
+    this.child = child;
+    this.state = state ?? LabelState(context);
+  }
 
   @override
-  List<Object?> get props => [
-        text,
-        isRequired,
-        enabled,
-      ];
+  LabelChild get child => super.child as LabelChild;
+
+  @override
+  LabelState get state => super.state as LabelState;
 }
 
+class LabelChild extends ChildImpl {
+  LabelChild(super.context, {required this.text});
+
+  final String text;
+}
+
+class LabelState extends StateImpl {
+  LabelState(super.context, {super.isDisabled, this.isRequired = false});
+
+  final bool isRequired;
+}
+
+typedef LabelDecorationBuilder = LabelDecoration Function(BuildContext context);
+
 class Label extends StatelessWidget {
-  final LabelModel vm;
-  const Label({super.key, required this.vm});
+  final LabelDecorationBuilder decorationBuilder;
+  const Label({super.key, required this.decorationBuilder});
 
   @override
   Widget build(BuildContext context) {
+    final decoration = decorationBuilder(context);
     return DefaultDisabled(
         decorationBuilder: (context) => DisabledDecoration(context,
-            state: DisabledState(context, isDisabled: !vm.enabled),
+            state:
+                DisabledState(context, isDisabled: decoration.state.isDisabled),
             child: DisabledChild(context,
                 child: RichText(
                     text: TextSpan(children: [
                   TextSpan(
-                      text: vm.text,
+                      text: decoration.child.text,
                       style: Theme.of(context).textTheme.titleSmall!.copyWith(
                             fontWeight: FontWeight.normal,
-                            color: vm.enabled
-                                ? null
-                                : Theme.of(context)
-                                    .colorScheme
-                                    .onSurface
-                                    .withOpacity(0.4),
+                            color: Theme.of(context)
+                                .colorScheme
+                                .onSurface
+                                .withOpacity(0.4),
                           )),
-                  if (vm.isRequired)
+                  if (decoration.state.isRequired)
                     TextSpan(
                         text: "\t*",
                         style: Theme.of(context)