+++ registry/lib/ui/default_components/radio.dart
@@ -2,84 +2,139 @@
 
 import 'package:fcnui_base/fcnui_base.dart';
 import 'package:flutter/material.dart';
+import 'package:registry/ui/default_components/fcnui_theme.dart';
 
 import 'disabled.dart';
 import 'form.dart';
 import 'dp_item.dart';
 
-class RadioModel extends IFormModel {
-  const RadioModel(
-      {required super.name,
-      required this.form,
-      this.decoration = const RadioDecoration()});
-
-  final RadioForm form;
-  final RadioDecoration decoration;
+class RadioDecoration extends DecorationImpl {
+  RadioDecoration(
+    super.context, {
+    RadioChild? child,
+    RadioColor? color,
+    required RadioValue value,
+    RadioAction? action,
+  }) {
+    super.value = value;
+    super.child = child ?? RadioChild(context);
+    super.color = color ?? RadioColor(context);
+    super.action = action ?? RadioAction(context);
+  }
 
   @override
-  List<Object?> get props => [name, form, decoration];
+  RadioChild get child => super.child as RadioChild;
+
+  @override
+  RadioColor get color => super.color as RadioColor;
+
+  @override
+  RadioValue get value => super.value as RadioValue;
+
+  @override
+  RadioAction get action => super.action as RadioAction;
 }
 
-class RadioForm extends Equatable {
+class RadioValue extends ValueImpl {
+  RadioValue(
+    super.context, {
+    required this.name,
+    required this.items,
+    this.validator,
+    this.initialValue,
+    this.disabledItems = const [],
+    this.autovalidateMode = AutovalidateMode.onUserInteraction,
+  });
+
+  final String name;
   final List<DpItem> items;
-  final ValueChanged<String?>? onChanged;
   final FormFieldValidator<String>? validator;
   final String? initialValue;
-  final List<String> disabled;
-  final AutovalidateMode? autovalidateMode;
-
-  const RadioForm(
-      {required this.items,
-      this.onChanged,
-      this.validator,
-      this.initialValue,
-      this.autovalidateMode,
-      this.disabled = const []});
-
-  @override
-  List<Object?> get props =>
-      [items, onChanged, validator, initialValue, disabled, autovalidateMode];
+  final List<String> disabledItems;
+  final AutovalidateMode autovalidateMode;
 }
 
-class RadioDecoration extends Equatable {
+class RadioChild extends ChildImpl {
+  RadioChild(
+    super.context, {
+    this.title,
+    this.controlAffinity = ControlAffinity.leading,
+    this.direction = OptionsOrientation.vertical,
+    this.separatorWidget,
+    TextStyle? titleStyle,
+    TextStyle? itemTitleStyle,
+    TextStyle? itemSubtitleStyle,
+  }) {
+    void setTextStyle() {
+      this.titleStyle = titleStyle ?? theme.textTheme.titleSmall;
+      this.itemTitleStyle = itemTitleStyle ?? theme.textTheme.bodyMedium;
+      this.itemSubtitleStyle = itemSubtitleStyle ??
+          theme.textTheme.bodySmall!
+              .copyWith(color: theme.colorScheme.onSurface.withOpacity(0.6));
+    }
+
+    setTextStyle();
+  }
+
   final String? title;
+  TextStyle? titleStyle;
   final ControlAffinity controlAffinity;
   final OptionsOrientation direction;
   final Widget? separatorWidget;
-  final Color? activeColor;
-  final Color? inactiveColor;
 
-  const RadioDecoration(
-      {this.title,
-      this.controlAffinity = ControlAffinity.leading,
-      this.direction = OptionsOrientation.vertical,
-      this.separatorWidget,
-      this.activeColor,
-      this.inactiveColor});
-
-  @override
-  List<Object?> get props =>
-      [title, controlAffinity, direction, separatorWidget];
+  TextStyle? itemTitleStyle;
+  TextStyle? itemSubtitleStyle;
 }
 
+class RadioColor extends ColorImpl {
+  RadioColor(
+    super.context, {
+    Color? activeColor,
+    Color? inactiveColor,
+    Color? disabledColor,
+    Color? errorColor,
+  }) {
+    void setColor() {
+      this.activeColor = activeColor ?? primary;
+      this.inactiveColor = inactiveColor ?? onSurface.withOpacity(0.6);
+      this.errorColor = errorColor ?? Colors.red;
+    }
+
+    setColor();
+  }
+
+  Color? activeColor;
+  Color? inactiveColor;
+  Color? disabledColor;
+  Color? errorColor;
+}
+
+class RadioAction extends ActionImpl<String> {
+  RadioAction(super.context, {this.onChanged});
+
+  final ValueChanged<String?>? onChanged;
+}
+
+typedef RadioDecorationBuilder = RadioDecoration Function(BuildContext context);
+
 class DefaultRadio extends StatelessWidget {
-  final RadioModel vm;
+  final RadioDecorationBuilder decorationBuilder;
 
-  const DefaultRadio({super.key, required this.vm});
+  const DefaultRadio({super.key, required this.decorationBuilder});
 
-  RadioThemeData _getRadioTheme(ThemeData theme) {
+  RadioThemeData _getRadioTheme(RadioDecoration decoration) {
     return RadioThemeData(
       splashRadius: 8,
       materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
       fillColor: MaterialStateProperty.resolveWith((states) {
         //if selected
         if (states.contains(MaterialState.selected)) {
-          return vm.decoration.activeColor;
+          return decoration.color.activeColor;
         }
         if (states.contains(MaterialState.disabled)) {
-          return null;
+          return decoration.color.disabledColor;
         }
-        return vm.decoration.inactiveColor;
+        return decoration.color.inactiveColor;
       }),
       mouseCursor: MaterialStateProperty.resolveWith((states) {
         //if disabled return not-allowed
@@ -93,16 +148,17 @@ class DefaultRadio extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final decoration = decorationBuilder(context);
     return DefaultDisabled(
         decorationBuilder: (context) => DisabledDecoration(context,
-            state:
-                DisabledState(context, isDisabled: vm.form.onChanged == null),
+            state: DisabledState(context,
+                isDisabled: decoration.action.onChanged == null),
             child: DisabledChild(context,
                 child: ThemeProvider(builder: (context, themeVm) {
               final theme = themeVm.theme;
               return Theme(
                   data: theme.copyWith(
-                    inputDecorationTheme: const InputDecorationTheme(
+                    inputDecorationTheme: InputDecorationTheme(
                         border: InputBorder.none,
                         focusedBorder: InputBorder.none,
                         enabledBorder: InputBorder.none,
@@ -110,44 +166,44 @@ class DefaultRadio extends StatelessWidget {
                         disabledBorder: InputBorder.none,
                         focusedErrorBorder: InputBorder.none,
                         contentPadding: EdgeInsets.zero,
-                        errorStyle: TextStyle(
-                          color: Colors.red,
-                        ),
+                        errorStyle:
+                            TextStyle(color: decoration.color.errorColor),
                         errorMaxLines: 2),
-                    radioTheme: _getRadioTheme(theme),
+                    radioTheme: _getRadioTheme(decoration),
                   ),
                   child: Column(
                       mainAxisSize: MainAxisSize.min,
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children: [
-                        if (vm.decoration.title != null)
-                          Text(vm.decoration.title!,
-                              style: theme.textTheme.titleSmall),
+                        if (decoration.child.title != null)
+                          Text(decoration.child.title!,
+                              style: decoration.child.titleStyle),
                         FormBuilderRadioGroup<String>(
-                            name: vm.name,
-                            onChanged: vm.form.onChanged,
-                            enabled: vm.form.onChanged != null,
-                            validator: vm.form.validator,
-                            initialValue: vm.form.initialValue,
-                            disabled: vm.form.disabled,
-                            controlAffinity: vm.decoration.controlAffinity,
-                            autovalidateMode: vm.form.autovalidateMode,
-                            separator: vm.decoration.separatorWidget,
-                            orientation: vm.decoration.direction,
-                            wrapDirection: vm.decoration.direction ==
+                            name: decoration.value.name,
+                            onChanged: decoration.action.onChanged,
+                            enabled: decoration.action.onChanged != null,
+                            validator: decoration.value.validator,
+                            initialValue: decoration.value.initialValue,
+                            disabled: decoration.value.disabledItems,
+                            controlAffinity: decoration.child.controlAffinity,
+                            autovalidateMode: decoration.value.autovalidateMode,
+                            separator: decoration.child.separatorWidget,
+                            orientation: decoration.child.direction,
+                            wrapDirection: decoration.child.direction ==
                                     OptionsOrientation.vertical
                                 ? Axis.vertical
                                 : Axis.horizontal,
                             wrapRunSpacing: 12.w,
                             wrapSpacing: 12.w,
-                            options: vm.form.items
+                            options: decoration.value.items
                                 .map((e) => FormBuilderFieldOption(
                                     value: e.id,
                                     child: DefaultDisabled(
                                         decorationBuilder: (context) =>
                                             DisabledDecoration(context,
                                                 state: DisabledState(context,
-                                                    isDisabled: vm.form.disabled
+                                                    isDisabled: decoration
+                                                        .value.disabledItems
                                                         .contains(e.id)),
                                                 child: DisabledChild(context,
                                                     child: Column(
@@ -158,20 +214,15 @@ class DefaultRadio extends StatelessWidget {
                                                                 .start,
                                                         children: [
                                                           Text(e.title,
-                                                              style: theme
-                                                                  .textTheme
-                                                                  .bodyMedium!),
+                                                              style: decoration
+                                                                  .child
+                                                                  .itemTitleStyle),
                                                           if (e.subtitle !=
                                                               null)
                                                             Text(e.subtitle!,
-                                                                style: theme
-                                                                    .textTheme
-                                                                    .bodySmall!
-                                                                    .copyWith(
-                                                                        color: theme
-                                                                            .colorScheme
-                                                                            .onSurface
-                                                                            .withOpacity(0.6)))
+                                                                style: decoration
+                                                                    .child
+                                                                    .itemSubtitleStyle)
                                                         ]))))))
                                 .toList())
                       ]).spaced(8));